Moved ICanLoad to view model.
