+++ src/GitHub.App/SampleData/PullRequestListViewModelDesigner.cs
@@ -9,6 +9,7 @@ using System.Collections.Generic;
 using ReactiveUI;
 using System.Collections.ObjectModel;
 using System.Linq;
+using GitHub.UI;
 
 namespace GitHub.SampleData
 {
@@ -56,7 +57,6 @@ namespace GitHub.SampleData
 
         public ITrackingCollection<IPullRequestModel> PullRequests { get; set; }
         public IPullRequestModel SelectedPullRequest { get; set; }
-        public ICommand OpenPullRequest { get; set; }
 
         public IReadOnlyList<PullRequestState> States { get; set; }
         public PullRequestState SelectedState { get; set; }
@@ -66,5 +66,9 @@ namespace GitHub.SampleData
 
         public ObservableCollection<IAccount> Assignees { get; set; }
         public IAccount SelectedAssignee { get; set; }
+        public IObservable<ViewWithData> Load { get; }
+
+        public ReactiveCommand<object> OpenPullRequest { get; }
+        public ReactiveCommand<object> CreatePullRequest { get; }
     }
 }
