+++ src/GitHub.Exports.Reactive/ViewModels/IPullRequestListViewModel.cs
@@ -27,16 +27,17 @@ namespace GitHub.ViewModels
         }
     }
 
-    public interface IPullRequestListViewModel : IViewModel
+    public interface IPullRequestListViewModel : IViewModel, ICanLoad
     {
         ITrackingCollection<IPullRequestModel> PullRequests { get; }
         IPullRequestModel SelectedPullRequest { get; }
-        ICommand OpenPullRequest { get; }
         IReadOnlyList<PullRequestState> States { get; set; }
         PullRequestState SelectedState { get; set; }
         ObservableCollection<IAccount> Authors { get; }
         IAccount SelectedAuthor { get; set; }
         ObservableCollection<IAccount> Assignees { get; }
         IAccount SelectedAssignee { get; set; }
+        ReactiveCommand<object> OpenPullRequest { get; }
+        ReactiveCommand<object> CreatePullRequest { get; }
     }
 }
